DateUpdate {lubridate}	R Documentation

Changes the components of a date object

Description

update.Date and update.POSIXt return a date with the specified elements updated. Elements not specified will be left unaltered. update.Date and update.POSIXt do not add the specified values to the existing date, they substitute them for the appropriate parts of the existing date.

Arguments

object
a date-time object

years
a value to substitute for the date's year component

months
a value to substitute for the date's month component

ydays
a value to substitute for the date's yday component

wdays
a value to substitute for the date's wday component

mdays
a value to substitute for the date's mday component

days
a value to substitute for the date's mday component

hours
a value to substitute for the date's hour component

minutes
a value to substitute for the date's minute component

seconds
a value to substitute for the date's second component

tzs
a value to substitute for the date's tz component

...
...

Value

a date object with the requested elements updated. The object will retain its original class unless an element is updated which the original class does not support. In this case, the date returned will be a POSIXlt date object.

Examples

date <- as.POSIXlt("2009-02-10")
update(date, year = 2010, month = 1, mday = 1)
# "2010-01-01 CST"

update(date, year =2010, month = 13, mday = 1)
# "2011-01-01 CST"

update(date, minute = 10, second = 3)
# "2009-02-10 00:10:03 CST"

[Package lubridate version 1.3.3 Index]