add_epoch_to_date {lubridate}	R Documentation

Add epochs to dates

Description

Method for adding epochs to dates. Epochs will be implemented in a later version of lubridate.

Usage

  add_epoch_to_date(date, epoch)

Arguments

date
a date-time object to be used as the reference time from which future epochs will be counted

epoch
an epoch object that decribes the number and type of epochs to be used to define a new date-time

Value

the date-time that occurs that specified number of epochs after the original date-time

[Package lubridate version 1.3.3 Index]