am {lubridate}	R Documentation

Does date time occur in the am or pm?

Description

Does date time occur in the am or pm?

Usage

  am(x)

Arguments

x
a date-time object

Value

TRUE or FALSE depending on whethe x occurs in the am or pm

Examples

x <- ymd("2012-03-26")
am(x)
pm(x)

[Package lubridate version 1.3.3 Index]