ceiling_date {lubridate}	R Documentation

Round date-times up.

Description

ceiling_date takes a date-time object and rounds it up to the nearest integer value of the specified time unit. Users can specify whether to round up to the nearest second, minute, hour, day, week, month, or year.

Usage

  ceiling_date(x,
    unit = c("second", "minute", "hour", "day", "week", "month", "year"))

Arguments

x
a vector of date-time objects

unit
a character string specifying the time unit to be rounded to. Should be one of "second","minute","hour","day", "week", "month", or "year."

Value

x with the appropriate units rounded up

See Also

floor_date, round_date

Examples

x <- as.POSIXct("2009-08-03 12:01:59.23")
ceiling_date(x, "second")
# "2009-08-03 12:02:00 CDT"
ceiling_date(x, "minute")
# "2009-08-03 12:02:00 CDT"
ceiling_date(x, "hour")
# "2009-08-03 13:00:00 CDT"
ceiling_date(x, "day")
# "2009-08-04 CDT"
ceiling_date(x, "week")
# "2009-08-09 CDT"
ceiling_date(x, "month")
# "2009-09-01 CDT"
ceiling_date(x, "year")
# "2010-01-01 CST"

[Package lubridate version 1.3.3 Index]