floor_date {lubridate}	R Documentation

Round date-times down.

Description

floor_date takes a date-time object and rounds it down to the nearest integer value of the specified time unit. Users can specify whether to round down to the nearest second, minute, hour, day, week, month, or year.

Usage

  floor_date(x,
    unit = c("second", "minute", "hour", "day", "week", "month", "year"))

Arguments

x
a vector of date-time objects

unit
a character string specifying the time unit to be rounded to. Should be one of "second","minute","hour","day", "week", "month", or "year."

Value

x with the appropriate units floored

See Also

ceiling_date, round_date

Examples

x <- as.POSIXct("2009-08-03 12:01:59.23")
floor_date(x, "second")
# "2009-08-03 12:01:59 CDT"
floor_date(x, "minute")
# "2009-08-03 12:01:00 CDT"
floor_date(x, "hour")
# "2009-08-03 12:00:00 CDT"
floor_date(x, "day")
# "2009-08-03 CDT"
floor_date(x, "week")
# "2009-08-02 CDT"
floor_date(x, "month")
# "2009-08-01 CDT"
floor_date(x, "year")
# "2009-01-01 CST"

[Package lubridate version 1.3.3 Index]