int_end {lubridate}	R Documentation

Access and change the end date of an interval

Description

Note that changing the end date of an interval will change the length of the interval, since the start date will remain the same.

Usage

  int_end(int)

Arguments

int
An interval object

Value

A POSIXct date object when used as an accessor. Nothing when used as a settor

See Also

int_start, int_shift, int_flip, int_length

Examples

int <- new_interval(ymd("2001-01-01"), ymd("2002-01-01"))
# 2001-01-01 UTC--2002-01-01 UTC
int_end(int)
# "2002-01-01 UTC"
int_end(int) <- ymd("2002-06-01")
int
# 2001-01-01 UTC--2002-06-01 UTC

[Package lubridate version 1.3.3 Index]