make_difftime {lubridate}	R Documentation

Makes a difftime object from a given number of seconds

Description

Makes a difftime object from a given number of seconds

Usage

  make_difftime(x)

Arguments

x
number value of seconds to be transformed into a difftime object

Value

a difftime object corresponding to x seconds

Examples

make_difftime(1)
make_difftime(60)
make_difftime(3600)

[Package lubridate version 1.3.3 Index]