minute {lubridate}	R Documentation

Get/set minutes component of a date-time.

Description

Date-time must be a POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.

Usage

  minute(x)

Arguments

x
a date-time object

Value

the minutes element of x as a decimal number

Examples

x <- ymd("2012-03-26")
minute(x)
minute(x) <- 1
minute(x) <- 61
minute(x) > 2

[Package lubridate version 1.3.3 Index]