month {lubridate}	R Documentation

Get/set months component of a date-time.

Description

Date-time must be a POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.

Usage

  month(x, label = FALSE, abbr = TRUE)

Arguments

x
a date-time object

label
logical. TRUE will display the month as a character string such as "January." FALSE will display the month as a number.

abbr
logical. FALSE will display the month as a character string #' label, such as #' "January". TRUE will display an abbreviated version of the label, such as "Jan". abbr is #' disregarded if label = FALSE.

Value

the months element of x as a number (1-12) or character string. 1 = January.

Examples

x <- ymd("2012-03-26")
month(x)
month(x) <- 1
month(x) <- 13
month(x) > 3

month(ymd(080101))
# 1
month(ymd(080101), label = TRUE)
# "January"
month(ymd(080101), label = TRUE, abbr = TRUE)
# "Jan"
month(ymd(080101) + months(0:11), label = TRUE, abbr = TRUE)
# "Jan" "Feb" "Mar" "Apr" "May" "Jun" "Jul" "Aug" "Sep" "Oct" "Nov" "Dec"

[Package lubridate version 1.3.3 Index]