new_epoch {lubridate}	R Documentation

Create an epoch object

Description

new_epoch creates an epoch object with the specified values. Implementation for epochs will be implemented in future versions of lubridate.

Usage

  new_epoch(weekday, number)

Arguments

weekday
name of weekday to use as epoch

number
number of epochs to include

Value

an epoch object

See Also

saturdays

[Package lubridate version 1.3.3 Index]