now {lubridate}	R Documentation

The current time

Description

The current time

Usage

now(tzone = "")

Arguments

tzone
a character vector specifying which time zone you would like the current time in. tzone defaults to your computer's system timezone. You can retrieve the current time in the Universal Coordinated Time (UTC) with now("UTC").

Value

the current date and time as a POSIXct object

See Also

here

Examples

now()
now("GMT")
now("")
now() == now() # would be true if computer processed both at the same instant
now() < now() # TRUE
now() > now() # FALSE

[Package lubridate version 1.6.0 Index]