olson_time_zones {lubridate}	R Documentation

Names of available time zones

Description

The names of all available Olson-style time zones.

Usage

  olson_time_zones(order_by = c("name", "longitude"))

Arguments

order_by
Return names alphabetically (the default) or from West to East.

Value

A character vector of time zone names.

Note

Olson-style names are the most readable and portable way of specifying time zones. This function gets names from the file shipped with R, stored in the file 'zone.tab'. ?Sys.timezone has more information.

See Also

Sys.timezone

Examples

## Not run: 
olson_time_zones()
olson_time_zones("longitude")

## End(Not run)

[Package lubridate version 1.3.3 Index]