period_to_seconds {lubridate}	R Documentation

Convert a period to the number of seconds it appears to represent

Description

Convert a period to the number of seconds it appears to represent

Usage

  period_to_seconds(x)

Arguments

x
A period object

[Package lubridate version 1.3.3 Index]