second {lubridate}	R Documentation

Get/set seconds component of a date-time

Description

Date-time must be a POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.

Usage

second(x)

second(x) <- value

Arguments

x
a date-time object

value
numeric value to be assigned

Value

the seconds element of x as a decimal number

Examples

x <- ymd("2012-03-26")
second(x)
second(x) <- 1
second(x) <- 61
second(x) > 2

[Package lubridate version 1.8.0 Index]