seconds_to_period {lubridate}	R Documentation

Contrive a period from a given number of seconds

Description

seconds_to_period uses estimates of time elements (in seconds) to create the period that has the maximum number of large elements(years > months > days > hours > minutes > seconds) and roughly equates to a given number of seconds. Note that the actual number of seconds in a period depends on when the period occurs. Since there is no one-to-one relationship between the periods that seconds_to_period makes and the number of seconds given as input, these periods should be treated as rough estimates only.

Usage

  seconds_to_period(x)

Arguments

x
A numeric object. The number of seconds to coerce into a period.

Value

A period that roughly equates to the number of seconds given.

[Package lubridate version 1.3.3 Index]