sundays {lubridate}	R Documentation

Quickly create common epoch objects

Description

Creates an epoch object that uses the functions name as the reference epoch event. Implementation for epochs will be implemented in future versions of lubridate.

Usage

  sundays(x = 1)

Arguments

x
number of epochs to be included

[Package lubridate version 1.3.3 Index]