today {lubridate}	R Documentation

The current date

Description

The current date

Usage

  today(tzone = "")

Arguments

tzone
a character vector specifying which time zone you would like to find the current date of. tzone defaults to the system time zone set on your computer.

Value

the current date as a Date object

Examples

today()
today("GMT")
today() == today("GMT") # not always true
today() < as.Date("2999-01-01") # TRUE  (so far)

[Package lubridate version 1.3.3 Index]