week {lubridate}	R Documentation

Get/set weeks component of a date-time.

Description

Date-time must be a POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects. Weeks is the number of complete seven day periods that have occured between the date and January 1st, plus one. isoweek returns the week as it would appear in the ISO 8601 system, which uses a reoccuring leap week.

Usage

  week(x)

Arguments

x
a date-time object

Value

the weeks element of x as an integer number

Examples

x <- ymd("2012-03-26")
week(x)
week(x) <- 1
week(x) <- 54
week(x) > 3

[Package lubridate version 1.3.3 Index]