year {lubridate}	R Documentation

Get/set years component of a date-time.

Description

Date-time must be a POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.

Usage

  year(x)

Arguments

x
a date-time object

Details

year does not yet support years before 0 C.E.

Value

the years element of x as a decimal number

Examples

x <- ymd("2012-03-26")
year(x)
year(x) <- 2001
year(x) > 1995

[Package lubridate version 1.3.3 Index]