date_decimal {lubridate}	R Documentation

Converts a decimal to a date

Description

Converts a decimal to a date

Usage

date_decimal(decimal, tz = "UTC")

Arguments

decimal
a numeric object

tz
the time zone required

Value

a POSIXct object, whose year corresponds to the integer part of decimal. The months, days, hours, minutes and seconds elements are picked so the date-time will accurately represent the fraction of the year expressed by decimal.

Examples

date <- ymd("2009-02-10")
decimal <- decimal_date(date)  # 2009.11
date_decimal(decimal) # "2009-02-10 UTC"

[Package lubridate version 1.8.0 Index]