day {lubridate}	R Documentation

Get/set days component of a date-time.

Description

Get/set days component of a date-time.

Usage

day(x)

mday(x)

wday(x, label = FALSE, abbr = TRUE)

qday(x)

yday(x)

day(x) <- value

mday(x) <- value

qday(x) <- value

wday(x) <- value

yday(x) <- value

Arguments

x
a POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, or fts object.

label
logical. Only available for wday. TRUE will display the day of the week as an ordered factor of character strings, such as "Sunday." FALSE will display the day of the week as a number.

abbr
logical. Only available for wday. FALSE will display the day of the week as an ordered factor of character strings, such as "Sunday." TRUE will display an abbreviated version of the label, such as "Sun". abbr is disregarded if label = FALSE.

value
a numeric object

Details

day and day<- are aliases for mday and mday<- respectively.

Value

wday returns the day of the week as a decimal number (01-07, Sunday is 1) or an ordered factor (Sunday is first).

See Also

yday, mday

Examples

x <- as.Date("2009-09-02")
wday(x) #4

wday(ymd(080101))
wday(ymd(080101), label = TRUE, abbr = FALSE)
# Levels: Sunday < Monday < Tuesday < Wednesday < Thursday < Friday < Saturday
wday(ymd(080101), label = TRUE, abbr = TRUE)
# Levels: Sunday < Monday < Tuesday < Wednesday < Thursday < Friday < Saturday
wday(ymd(080101) + days(-2:4), label = TRUE, abbr = TRUE)
# Levels: Sunday < Monday < Tuesday < Wednesday < Thursday < Friday < Saturday

x <- as.Date("2009-09-02")
yday(x) #245
mday(x) #2
yday(x) <- 1  #"2009-01-01"
yday(x) <- 366 #"2010-01-01"
mday(x) > 3

[Package lubridate version 1.6.0 Index]