days_in_month {lubridate}	R Documentation

Get the number of days in the month of a date-time

Description

Date-time must be a POSIXct, POSIXlt, Date, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.

Usage

days_in_month(x)

Arguments

x
a date-time object

Value

An integer of the number of days in the month component of the date-time object.

[Package lubridate version 1.8.0 Index]