hour {lubridate}	R Documentation

Get/set hours component of a date-time

Description

Date-time must be a POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr, zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts objects.

Usage

hour(x)

hour(x) <- value

Arguments

x
a date-time object

value
numeric value to be assigned to the hour component

Value

the hours element of x as a decimal number

Examples

x <- ymd("2012-03-26")
hour(x)
hour(x) <- 1
hour(x) <- 25
hour(x) > 2

[Package lubridate version 1.8.0 Index]