lakers {lubridate}	R Documentation

Lakers 2008-2009 basketball data set

Description

This data set contains play by play statistics of each Los Angeles Lakers basketball game in the 2008-2009 season. Data includes the date, opponent, and type of each game (home or away). Each play is described by the time on the game clock when the play was made, the period in which the play was attempted, the type of play, the player and team who made the play, the result of the play, and the location on the court where each play was made.

References

Originally taken from www.basketballgeek.com/data/.

[Package lubridate version 1.8.0 Index]