local_time {lubridate}	R Documentation

Get local time from a date-time vector.

Description

local_time retrieves day clock time in specified time zones. Computation is vectorized over both dt and tz arguments, the shortest is recycled in accordance with standard R rules.

Usage

local_time(dt, tz = NULL, units = "secs")

Arguments

dt
a date-time object.

tz
a character vector of timezones for which to compute the local time.

units
passed directly to as.difftime().

Examples

x <- ymd_hms(c("2009-08-07 01:02:03", "2009-08-07 10:20:30"))
local_time(x, units = "secs")
local_time(x, units = "hours")
local_time(x, "Europe/Amsterdam")
local_time(x, "Europe/Amsterdam") == local_time(with_tz(x, "Europe/Amsterdam"))

x <- ymd_hms("2009-08-07 01:02:03")
local_time(x, c("America/New_York", "Europe/Amsterdam", "Asia/Shanghai"), unit = "hours")

[Package lubridate version 1.9.3 Index]