now {lubridate}	R Documentation

The current day and time

Description

The current day and time

Usage

now(tzone = "")

today(tzone = "")

Arguments

tzone
a character vector specifying which time zone you would like the current time in. tzone defaults to your computer's system timezone. You can retrieve the current time in the Universal Coordinated Time (UTC) with now("UTC").

Value

now - the current datetime as a POSIXct object

Examples

now()
now("GMT")
now("")
now() == now() # would be TRUE if computer processed both at the same instant
now() < now() # TRUE
now() > now() # FALSE
today()
today("GMT")
today() == today("GMT") # not always true
today() < as.Date("2999-01-01") # TRUE  (so far)

[Package lubridate version 1.9.3 Index]