pretty_dates {lubridate}	R Documentation

Computes attractive axis breaks for date-time data

Description

pretty.dates indentifies which unit of time the sub-intervals should be measured in to provide approximately n breaks. It then chooses a "pretty" length for the sub-intervals and sets start and endpoints that 1) span the entire range of the data, and 2) allow the breaks to occur on important date-times (i.e. on the hour, on the first of the month, etc.)

Usage

pretty_dates(x, n, ...)

Arguments

x
a vector of POSIXct, POSIXlt, Date, or chron date-time objects

n
integer value of the desired number of breaks

...
additional arguments to pass to function

Value

a vector of date-times that can be used as axis tick marks or bin breaks

Examples

x <- seq.Date(as.Date("2009-08-02"), by = "year", length.out = 2)
pretty_dates(x, 12)

[Package lubridate version 1.9.3 Index]