time_length {lubridate}	R Documentation

Compute the exact length of a time span

Description

Compute the exact length of a time span

Usage

time_length(x, unit = "second")

## S4 method for signature 'Interval'
time_length(x, unit = "second")

Arguments

x
a duration, period, difftime or interval

unit
a character string that specifies with time units to use

Details

When x is an Interval object and unit are years or months, time_length() takes into account the fact that all months and years don't have the same number of days.

When x is a Duration, Period or difftime() object, length in months or years is based on their most common lengths in seconds (see timespan()).

Value

the length of the interval in the specified unit. A negative number connotes a negative interval or duration

See Also

timespan()

Examples

int <- interval(ymd("1980-01-01"), ymd("2014-09-18"))
time_length(int, "week")

# Exact age
time_length(int, "year")

# Age at last anniversary
trunc(time_length(int, "year"))

# Example of difference between intervals and durations
int <- interval(ymd("1900-01-01"), ymd("1999-12-31"))
time_length(int, "year")
time_length(as.duration(int), "year")

[Package lubridate version 1.8.0 Index]