%within% {lubridate}	R Documentation

Tests whether a date or interval falls within an interval

Description

If a is an interval, both its start and end dates must fall within b to return TRUE.

Usage

a %within% b

Arguments

a
An interval or date-time object

b
An interval

Value

A logical

Examples

int <- interval(ymd("2001-01-01"), ymd("2002-01-01"))
int2 <- interval(ymd("2001-06-01"), ymd("2002-01-01"))

ymd("2001-05-03") %within% int # TRUE
int2 %within% int # TRUE
ymd("1999-01-01") %within% int # FALSE

[Package lubridate version 1.6.0 Index]