year {lubridate}	R Documentation

Get/set years component of a date-time

Description

Date-time must be a POSIXct, POSIXlt, Date, Period or any other object convertible to POSIXlt.

isoyear() returns years according to the ISO 8601 week calendar.

epiyear() returns years according to the epidemilogical week calendars.

Usage

year(x)

year(x) <- value

isoyear(x)

epiyear(x)

Arguments

x
a date-time object

value
a numeric object

Details

year does not yet support years before 0 C.E.

Value

the years element of x as a decimal number

References

https://en.wikipedia.org/wiki/ISO_week_date

Examples

x <- ymd("2012-03-26")
year(x)
year(x) <- 2001
year(x) > 1995

[Package lubridate version 1.8.0 Index]